refactor: 예외 추가
