+++ src/main/java/subway/exception/LineDuplicatedException.java
@@ -1,4 +1,4 @@
-package subway2.exception;
+package subway.exception;
 
 public class LineDuplicatedException extends RuntimeException {
