+++ src/main/java/subway/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package subway2.application;
+package subway.exception;
 
 public class NotFoundException extends RuntimeException {
