+++ src/main/java/subway/ui/GlobalExceptionHandler.java
@@ -1,14 +1,14 @@
-package subway2.ui;
+package subway.ui;
 
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import subway2.application.NotFoundException;
-import subway2.dto.ExceptionResponse;
-import subway2.exception.InvalidSectionException;
-import subway2.exception.LineDuplicatedException;
+import subway.exception.InvalidSectionException;
+import subway.exception.LineDuplicatedException;
+import subway.exception.NotFoundException;
+import subway.ui.dto.ExceptionResponse;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
