+++ src/main/java/subway/ui/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package subway2.dto;
+package subway.ui.dto;
 
 import java.util.List;
